+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Complete configuration hierarchy bound (root, all classes between root and Configuration and Configuration itself)
  - Interfaces directly implemented by classes in configuration hierarchy expect interfaces from java package 
  (it's common to use HasSomeConfig interface convention and now interface may be directly used for binding)
+ - Add bindConfigurationInterfaces() option to GuiceBundle to be able to disable interface bindings
 
 ### 3.2.0 (2016-01-23)
 * Clear possible duplicate guicey bundle instances